requirements.txt
# Core dependencies
opencv-contrib-python>=4.5.5
numpy>=2.0.0
insightface>=0.7.3
onnxruntime>=1.15.0

# Backend Framework
Flask>=2.0.1
Flask-Cors>=3.0.10
flask-restx>=1.1.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
python-dotenv>=0.19.0

# Development
black>=21.0
flake8>=3.9.0
isort>=5.9.0
